+++ CHANGELOG.md
@@ -41,6 +41,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed ability to block yourself
 - Fixed show SSO login option with empty identity provider list
 - Don't show "Image could not be loaded" message while loading an image
+- Fixed jumping timeline on fast message sending
+- Fixed various issues, that unread marker is unnecessarily displayed
+- Fixed unnecessary re-computations on timeline re-init
+- Fixed typo in verification help text
 
 ### Security